requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
uuid
python-jose
passlib
